refactor : CompositionalLayout 메서드 확장을 통해 분리 #17
